select the dark vector boson from the mc collection
only copy this truth particle for the signal samples, this lightens the final analysis tuples and helps simplify the final analysis since we don't have to sift through all the MC particles to find the only one we care about.

Impose a requirement that one and exactly one true Vd is the in the event. This is necessary for decay re-weighting and the very rare events where there are zero or more than one Vd pale in comparison to the requirement of a reconstructed vertex